+++ palette_gen/solvers/impl/fixed.py
@@ -1,5 +1,5 @@
-from dataclasses import dataclass
 from collections.abc import Iterable
+from dataclasses import dataclass
 
 from matplotlib.colors import to_rgb