+++ palette_gen/util.py
@@ -1,5 +1,5 @@
-from typing import Any, TypeVar
 from collections.abc import Callable
+from typing import Any, TypeVar
 
 T = TypeVar("T")
 V = TypeVar("V")