+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0
+
+* (breaking) large refactor to make processors class based. This breaks the entrypoint.
+* (feat) add Neovim lua table processor
+
 ## 1.1
 
 * (proj) added changelog